+++ integration/templates/astro-node/src/env.d.ts
@@ -1,0 +1,3 @@
+/// <reference path="../.astro/types.d.ts" />
+/// <reference types="astro/client" />
+/// <reference types="@clerk/astro/env" />