+++ packages/astro/src/astro-components/interactive/index.ts
@@ -1,0 +1,18 @@
+// The `ts-ignore` comments here are necessary because we're importing this file inside the `astro:components`
+// virtual module's types, which means that `tsc` will try to resolve these imports. Don't mind the editor errors.
+// @ts-ignore
+export { default as SignIn } from './SignIn.astro';
+// @ts-ignore
+export { default as SignUp } from './SignUp.astro';
+// @ts-ignore
+export { default as UserButton } from './UserButton.astro';
+// @ts-ignore
+export { default as UserProfile } from './UserProfile.astro';
+// @ts-ignore
+export { default as OrganizationProfile } from './OrganizationProfile.astro';
+// @ts-ignore
+export { default as OrganizationSwitcher } from './OrganizationSwitcher.astro';
+// @ts-ignore
+export { default as OrganizationList } from './OrganizationList.astro';
+// @ts-ignore
+export { default as GoogleOneTap } from './GoogleOneTap.astro';