+++ packages/nextjs/src/server/clerkMiddleware.ts
@@ -1,3 +1,5 @@
+import { AsyncLocalStorage } from 'node:async_hooks';
+
 import type {
   AuthenticateRequestOptions,
   AuthObject,
@@ -69,6 +71,8 @@ interface ClerkMiddleware {
   (request: NextMiddlewareRequestParam, event: NextMiddlewareEvtParam): NextMiddlewareReturn;
 }
 
+export const clerkMiddlewareRequestDataStore = new AsyncLocalStorage<Partial<AuthenticateRequestOptions>>();
+
 export const clerkMiddleware: ClerkMiddleware = withLogger('clerkMiddleware', logger => (...args: unknown[]): any => {
   const [request, event] = parseRequestAndEvent(args);
   const [handler, params] = parseHandlerAndOptions(args);
@@ -91,82 +95,87 @@ export const clerkMiddleware: ClerkMiddleware = withLogger('clerkMiddleware', lo
     signUpUrl,
   };
 
-  clerkClient.telemetry.record(
-    eventMethodCalled('clerkMiddleware', {
-      handler: Boolean(handler),
-      satellite: Boolean(options.isSatellite),
-      proxy: Boolean(options.proxyUrl),
-    }),
-  );
-
-  const nextMiddleware: NextMiddleware = async (request, event) => {
-    const clerkRequest = createClerkRequest(request);
-    logger.debug('options', options);
-    logger.debug('url', () => clerkRequest.toJSON());
-
-    const requestState = await clerkClient.authenticateRequest(
-      clerkRequest,
-      createAuthenticateRequestOptions(clerkRequest, options),
+  return clerkMiddlewareRequestDataStore.run(options, () => {
+    clerkClient().telemetry.record(
+      eventMethodCalled('clerkMiddleware', {
+        handler: Boolean(handler),
+        satellite: Boolean(options.isSatellite),
+        proxy: Boolean(options.proxyUrl),
+      }),
     );
 
-    logger.debug('requestState', () => ({
-      status: requestState.status,
-      headers: JSON.stringify(Object.fromEntries(requestState.headers)),
-      reason: requestState.reason,
-    }));
+    const nextMiddleware: NextMiddleware = async (request, event) => {
+      const clerkRequest = createClerkRequest(request);
+      logger.debug('options', options);
+      logger.debug('url', () => clerkRequest.toJSON());
 
-    const locationHeader = requestState.headers.get(constants.Headers.Location);
-    if (locationHeader) {
-      return new Response(null, { status: 307, headers: requestState.headers });
-    } else if (requestState.status === AuthStatus.Handshake) {
-      throw new Error('Clerk: handshake status without redirect');
+      const requestState = await clerkClient().authenticateRequest(
+        clerkRequest,
+        createAuthenticateRequestOptions(clerkRequest, options),
+      );
+
+      logger.debug('requestState', () => ({
+        status: requestState.status,
+        headers: JSON.stringify(Object.fromEntries(requestState.headers)),
+        reason: requestState.reason,
+      }));
+
+      const locationHeader = requestState.headers.get(constants.Headers.Location);
+      if (locationHeader) {
+        return new Response(null, { status: 307, headers: requestState.headers });
+      } else if (requestState.status === AuthStatus.Handshake) {
+        throw new Error('Clerk: handshake status without redirect');
+      }
+
+      const authObject = requestState.toAuth();
+      logger.debug('auth', () => ({ auth: authObject, debug: authObject.debug() }));
+
+      const redirectToSignIn = createMiddlewareRedirectToSignIn(clerkRequest);
+      const protect = createMiddlewareProtect(clerkRequest, authObject, redirectToSignIn);
+      const authObjWithMethods: ClerkMiddlewareAuthObject = Object.assign(authObject, { protect, redirectToSignIn });
+
+      let handlerResult: Response = NextResponse.next();
+      try {
+        const userHandlerResult = await clerkMiddlewareRequestDataStore.run(options, async () =>
+          handler?.(() => authObjWithMethods, request, event),
+        );
+        handlerResult = userHandlerResult || handlerResult;
+      } catch (e: any) {
+        handlerResult = handleControlFlowErrors(e, clerkRequest, requestState);
+      }
+
+      // TODO @nikos: we need to make this more generic
+      // and move the logic in clerk/backend
+      if (requestState.headers) {
+        requestState.headers.forEach((value, key) => {
+          handlerResult.headers.append(key, value);
+        });
+      }
+
+      if (isRedirect(handlerResult)) {
+        logger.debug('handlerResult is redirect');
+        return serverRedirectWithAuth(clerkRequest, handlerResult, options);
+      }
+
+      if (options.debug) {
+        setRequestHeadersOnNextResponse(handlerResult, clerkRequest, { [constants.Headers.EnableDebug]: 'true' });
+      }
+
+      decorateRequest(clerkRequest, handlerResult, requestState, params);
+
+      return handlerResult;
+    };
+
+    // If we have a request and event, we're being called as a middleware directly
+    // eg, export default clerkMiddleware;
+    if (request && event) {
+      return nextMiddleware(request, event);
     }
 
-    const authObject = requestState.toAuth();
-    logger.debug('auth', () => ({ auth: authObject, debug: authObject.debug() }));
-
-    const redirectToSignIn = createMiddlewareRedirectToSignIn(clerkRequest);
-    const protect = createMiddlewareProtect(clerkRequest, authObject, redirectToSignIn);
-    const authObjWithMethods: ClerkMiddlewareAuthObject = Object.assign(authObject, { protect, redirectToSignIn });
-
-    let handlerResult: Response = NextResponse.next();
-    try {
-      handlerResult = (await handler?.(() => authObjWithMethods, request, event)) || handlerResult;
-    } catch (e: any) {
-      handlerResult = handleControlFlowErrors(e, clerkRequest, requestState);
-    }
-
-    // TODO @nikos: we need to make this more generic
-    // and move the logic in clerk/backend
-    if (requestState.headers) {
-      requestState.headers.forEach((value, key) => {
-        handlerResult.headers.append(key, value);
-      });
-    }
-
-    if (isRedirect(handlerResult)) {
-      logger.debug('handlerResult is redirect');
-      return serverRedirectWithAuth(clerkRequest, handlerResult, options);
-    }
-
-    if (options.debug) {
-      setRequestHeadersOnNextResponse(handlerResult, clerkRequest, { [constants.Headers.EnableDebug]: 'true' });
-    }
-
-    decorateRequest(clerkRequest, handlerResult, requestState, options.secretKey);
-
-    return handlerResult;
-  };
-
-  // If we have a request and event, we're being called as a middleware directly
-  // eg, export default clerkMiddleware;
-  if (request && event) {
-    return nextMiddleware(request, event);
-  }
-
-  // Otherwise, return a middleware that can be called with a request and event
-  // eg, export default clerkMiddleware(auth => { ... });
-  return nextMiddleware;
+    // Otherwise, return a middleware that can be called with a request and event
+    // eg, export default clerkMiddleware(auth => { ... });
+    return nextMiddleware;
+  });
 });
 
 const parseRequestAndEvent = (args: unknown[]) => {