+++ packages/tanstack-start/CHANGELOG.md
@@ -1,5 +1,29 @@
 # @clerk/tanstack-start
 
+## 0.1.5
+
+### Patch Changes
+
+- Updated dependencies [[`427fcdeaa`](https://github.com/clerk/javascript/commit/427fcdeaaba4e77273be29b4d7cca43f9aa18693)]:
+  - @clerk/clerk-react@5.2.7
+
+## 0.1.4
+
+### Patch Changes
+
+- Updated dependencies [[`f1847b70b`](https://github.com/clerk/javascript/commit/f1847b70b2327bd490faf1f3eed1aa5639d54993)]:
+  - @clerk/backend@1.3.0
+
+## 0.1.3
+
+### Patch Changes
+
+- Updated dependencies [[`d6b5006c4`](https://github.com/clerk/javascript/commit/d6b5006c4cc1b6f07bb3a6832b4ec6e65ea15814)]:
+  - @clerk/types@4.7.0
+  - @clerk/backend@1.2.5
+  - @clerk/clerk-react@5.2.6
+  - @clerk/shared@2.3.2
+
 ## 0.1.2
 
 ### Patch Changes